+++ vue/src/app/playground/index.vue
@@ -2,6 +2,7 @@
   <div class="row">
     <div class="example-spacing">
       <!-- Insert Example Here -->
+      <!--
       <div class="gap" />
       <ExampleDataTableDefault />
       <div class="gap" />
@@ -12,6 +13,22 @@
       <ExampleDataTableCustomSort />
       <div class="gap" />
       <ExampleDataTableResize />
+      <div class="gap" />
+      <ExampleDataTableZebra />
+      <div class="gap" />
+      <ExampleDataTableAvatar />
+      <div class="gap" />
+      <ExampleDataTableSelect />
+      <div class="gap" />
+      <ExampleDataTableResizeScroll />
+      <div class="gap" />
+      <ExampleDataTableSelection />
+      <div class="gap" />
+      <ExampleDataTableVirtualScroll />
+      -->
+      
+      <div class="gap" />
+      <ExampleDataTableMultiSelectCheckbox />
     </div>
   </div>
 </template>