+++ vue/src/lib/data-table/examples/index.js
@@ -1,16 +1,30 @@
 import DataTable from '../index.vue';
+import ExampleDataTableAvatar from './avatar.vue';
 import ExampleDataTableDefault from './default.vue';
+import ExampleDataTableMultiSelectCheckbox from './multi-select-checkbox.vue';
 import ExampleDataTableResize from './resize.vue';
+import ExampleDataTableResizeScroll from './resize-scroll.vue';
 import ExampleDataTableScroll from './scroll.vue';
+import ExampleDataTableSelect from './select.vue';
+import ExampleDataTableSelection from './selection.vue';
 import ExampleDataTableSort from './sort.vue';
 import ExampleDataTableCustomSort from './custom-sort.vue';
+import ExampleDataTableVirtualScroll from './virtual-scroll.vue';
+import ExampleDataTableZebra from './zebra.vue';
 
 export default {
   DataTable,
+  ExampleDataTableAvatar,
   ExampleDataTableDefault,
+  ExampleDataTableMultiSelectCheckbox,
   ExampleDataTableResize,
+  ExampleDataTableResizeScroll,
   ExampleDataTableScroll,
+  ExampleDataTableSelect,
+  ExampleDataTableSelection,
   ExampleDataTableSort,
   ExampleDataTableCustomSort,
+  ExampleDataTableVirtualScroll,
+  ExampleDataTableZebra
 };