+++ vue/src/lib/data-table/examples/sort.vue
@@ -7,7 +7,6 @@
 </template>
 
 <script>
-import moment from 'moment';
 export default {
   name: 'ExampleDataTableSort',